implement GetRandomTarget and MoveTowardsTarget behavior in inefficent way

+++ includes/config.h
@@ -24,5 +24,5 @@ namespace config {
     extern float enemySightRangeChasing;
     extern float enemyHearRange;
     extern float enemyHearRangeChasing;
-    extern int enemyPatrolDistance;
+    extern float enemyPatrolDistance;
 }
